done resolve and delete
