+++ demo/src/App.tsx
@@ -30,7 +30,7 @@ function App() {
   const [selectedTags, setSelectedTags] = useState<TagType[]>([]);
   const [isCommentSectionOpen, setIsCommentSectionOpen] = useState(false);
   const [isPresentationMode, setIsPresentationMode] = useState(false);
-  const [inlineCommentOpen, setInlineCommentOpen] = useState(false);
+  const [commentDrawerOpen, setCommentDrawerOpen] = useState(false);
   const [inlineCommentData, setInlineCommentData] = useState({
     inlineCommentText: '',
     highlightedTextContent: '',
@@ -87,7 +87,7 @@ function App() {
         </div>
         <div className="flex gap-2">
           <IconButton variant={'ghost'} icon="Presentation" size="md" onClick={() => setIsPresentationMode(true)} />
-          <IconButton variant={'ghost'} icon="MessageSquareText" size="md" onClick={() => setInlineCommentOpen(prev => !prev)} />
+          <IconButton variant={'ghost'} icon="MessageSquareText" size="md" onClick={() => setCommentDrawerOpen(prev => !prev)} />
           <IconButton
             variant={'ghost'}
             icon="Share2"
@@ -140,8 +140,8 @@ function App() {
         setIsCommentSectionOpen={setIsCommentSectionOpen}
         setInlineCommentData={setInlineCommentData}
         inlineCommentData={inlineCommentData}
-        inlineCommentOpen={inlineCommentOpen}
-        setInlineCommentOpen={setInlineCommentOpen}
+        commentDrawerOpen={commentDrawerOpen}
+        setCommentDrawerOpen={setCommentDrawerOpen}
         isPresentationMode={isPresentationMode}
         setIsPresentationMode={setIsPresentationMode}
         zoomLevel={zoomLevel}
