+++ package/types.ts
@@ -24,8 +24,8 @@ export type InlineCommentData = {
   handleClick: boolean;
 };
 export interface DdocProps {
-  inlineCommentOpen?: boolean;
-  setInlineCommentOpen?: React.Dispatch<SetStateAction<boolean>>;
+  commentDrawerOpen?: boolean;
+  setCommentDrawerOpen?: React.Dispatch<SetStateAction<boolean>>;
   selectedTags?: TagType[];
   setSelectedTags?: React.Dispatch<SetStateAction<TagType[]>>;
   enableCollaboration?: boolean | undefined;
